volume commands refactoring
Closes #35
