+++ commands/volumes/volume-delete.go
@@ -1,24 +1,17 @@
-// Package volumedelete implements the volume delete command for GlusterD
-package volumedelete
+package volumecommands
 
 import (
 	"net/http"
 
 	"github.com/gluster/glusterd2/client"
 	"github.com/gluster/glusterd2/errors"
-	"github.com/gluster/glusterd2/rest"
 	"github.com/gluster/glusterd2/volume"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/gorilla/mux"
 )
 
-// Command is a holding struct used to implement the GlusterD Command interface
-// for the volume delete command
-type Command struct {
-}
-
-func (c *Command) volumeDeleteHandler(w http.ResponseWriter, r *http.Request) {
+func volumeDeleteHandler(w http.ResponseWriter, r *http.Request) {
 	p := mux.Vars(r)
 	volname := p["volname"]
 
@@ -38,16 +31,4 @@ func (c *Command) volumeDeleteHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	client.SendResponse(w, 0, 0, "", http.StatusOK, "")
-}
-
-// Routes returns command routes to be set up for the volume delete command.
-func (c *Command) Routes() rest.Routes {
-	return rest.Routes{
-		// VolumeDelete
-		rest.Route{
-			Name:        "VolumeDelete",
-			Method:      "DELETE",
-			Pattern:     "/volumes/{volname}",
-			HandlerFunc: c.volumeDeleteHandler},
-	}
 }
