+++ commands/volumes/volume-info.go
@@ -1,0 +1,27 @@
+package volumecommands
+
+import (
+	"net/http"
+
+	"github.com/gluster/glusterd2/client"
+	"github.com/gluster/glusterd2/errors"
+	"github.com/gluster/glusterd2/volume"
+
+	log "github.com/Sirupsen/logrus"
+	"github.com/gorilla/mux"
+)
+
+func volumeInfoHandler(w http.ResponseWriter, r *http.Request) {
+	p := mux.Vars(r)
+	volname := p["volname"]
+
+	log.Debug("In Volume info API")
+
+	vol, e := volume.GetVolume(volname)
+	if e != nil {
+		client.SendResponse(w, -1, http.StatusNotFound, errors.ErrVolNotFound.Error(), http.StatusNotFound, "")
+	} else {
+
+		client.SendResponse(w, 0, 0, "", 0, vol)
+	}
+}
