+++ commands/volumes/volume-start.go
@@ -1,24 +1,17 @@
-// Package volumestart implements the volume start command for GlusterD
-package volumestart
+package volumecommands
 
 import (
 	"net/http"
 
 	"github.com/gluster/glusterd2/client"
 	"github.com/gluster/glusterd2/errors"
-	"github.com/gluster/glusterd2/rest"
 	"github.com/gluster/glusterd2/volume"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/gorilla/mux"
 )
 
-// Command is a holding struct used to implement the GlusterD Command interface
-// for the volume start command
-type Command struct {
-}
-
-func (c *Command) volumeStartHandler(w http.ResponseWriter, r *http.Request) {
+func volumeStartHandler(w http.ResponseWriter, r *http.Request) {
 	p := mux.Vars(r)
 	volname := p["volname"]
 
@@ -42,16 +35,4 @@ func (c *Command) volumeStartHandler(w http.ResponseWriter, r *http.Request) {
 	}
 	log.WithField("volume", vol.Name).Debug("Volume updated into the store")
 	client.SendResponse(w, 0, 0, "", http.StatusOK, vol)
-}
-
-// Routes returns command routes to be set up for the volume start command.
-func (c *Command) Routes() rest.Routes {
-	return rest.Routes{
-		// VolumeStart
-		rest.Route{
-			Name:        "VolumeStart",
-			Method:      "POST",
-			Pattern:     "/volumes/{volname}/start",
-			HandlerFunc: c.volumeStartHandler},
-	}
 }
